Initial changes and PoC
