+++ plugin.video.amazon-test/addon.xml
@@ -18,6 +18,7 @@
   </extension>
   <extension point="xbmc.addon.metadata">
     <platform>all</platform>
+    <reuselanguageinvoker>true</reuselanguageinvoker>
     <summary lang="en">Amazon Prime Video Streaming</summary>
     <summary lang="de">Amazon Prime Instant Video</summary>
     <description lang="en">Movies and Television Shows for Prime Members</description>
