+++ plugin.video.amazon-test/default.py
@@ -2,7 +2,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 from base64 import b64encode, b64decode
-from collections import OrderedDict
 from BeautifulSoup import BeautifulSoup, BeautifulStoneSoup, Tag
 from datetime import date
 from platform import node
@@ -37,7 +36,11 @@ import xbmcgui
 import xbmcplugin
 import xbmcvfs
 
-addon = xbmcaddon.Addon()
+from resources.lib import AmazonPlayer, PrimeVideo
+from resources.lib.common import Globals
+
+g = Globals()
+addon = g.addon
 Dialog = xbmcgui.Dialog()
 pDialog = xbmcgui.DialogProgress()
 pluginhandle = int(sys.argv[1])
@@ -98,27 +101,6 @@ AgesCfg = {'A1PA6795UKMFR9': ['FSK', ('FSK 0', 'FSK 0'), ('FSK 6', 'FSK 6'), ('F
            'ATVPDKIKX0DER': ['', ('General Audiences', 'G,TV-G,TV-Y'), ('Family', 'PG,NR,TV-Y7,TV-Y7-FV,TV-PG'), ('Teen', 'PG-13,TV-14'),
                               ('Mature', 'R,NC-17,TV-MA,Unrated,Not rated')],
            'A1VC38T7YXB528': ['', ('全ての観客', 'g'), ('親の指導・助言', 'pg12'), ('R-15指定', 'r15+'), ('成人映画', 'r18+,nr')]}
-
-dateParserData = {
-    'de_DE': {'deconstruct': r'^([0-9]+)\.\s+([^\s]+)\s+([0-9]+)', 'reassemble': '{2}-{1:0>2}-{0:0>2}', 'month': 1,
-              'months': {'Januar': 1, 'Februar': 2, 'März': 3, 'April': 4, 'Mai': 5, 'Juni': 6, 'Juli': 7, 'August': 8, 'September': 9, 'Oktober': 10,
-                         'November': 11, 'Dezember': 12}},
-    'en_US': {'deconstruct': r'^([^\s]+)\s+([0-9]+),\s+([0-9]+)', 'reassemble': '{2}-{0:0>2}-{1:0>2}', 'month': 0,
-              'months': {'January': 1, 'February': 2, 'March': 3, 'April': 4, 'May': 5, 'June': 6, 'July': 7, 'August': 8, 'September': 9, 'October': 10,
-                         'November': 11, 'December': 12}},
-    'es_ES': {'deconstruct': r'^([0-9]+)\s+de\s+([^\s]+),\s+de\s+([0-9]+)', 'reassemble': '{2}-{1:0>2}-{0:0>2}', 'month': 1,
-              'months': {'enero': 1, 'febrero': 2, 'marzo': 3, 'abril': 4, 'mayo': 5, 'junio': 6, 'julio': 7, 'agosto': 8, 'septiembre': 9, 'octubre': 10,
-                         'noviembre': 11, 'diciembre': 12}},
-    'fr_FR': {'deconstruct': r'^([0-9]+)\s+([^\s]+)\s+([0-9]+)', 'reassemble': '{2}-{1:0>2}-{0:0>2}', 'month': 1,
-              'months': {'janvier': 1, 'février': 2, 'mars': 3, 'avril': 4, 'mai': 5, 'juin': 6, 'juillet': 7, 'aout': 8, 'septembre': 9, 'octobre': 10,
-                         'novembre': 11, 'décembre': 12}},
-    'it_IT': {'deconstruct': r'^([0-9]+)\s+([^\s]+)\s+([0-9]+)', 'reassemble': '{2}-{1:0>2}-{0:0>2}', 'month': 1,
-              'months': {'gennaio': 1, 'febbraio': 2, 'marzo': 3, 'aprile': 4, 'maggio': 5, 'giugno': 6, 'luglio': 7, 'agosto': 8, 'settembre': 9,
-                         'ottobre': 10, 'novembre': 11, 'dicembre': 12}},
-    'pt_BR': {'deconstruct': r'^([0-9]+)\s+de\s+([^\s]+),\s+de\s+([0-9]+)', 'reassemble': '{2}-{1:0>2}-{0:0>2}', 'month': 1,
-              'months': {'janeiro': 1, 'fevereiro': 2, 'março': 3, 'abril': 4, 'maio': 5, 'junho': 6, 'julho': 7, 'agosto': 8, 'setembro': 9, 'outubro': 10,
-                         'novembro': 11, 'dezembro': 12}},
-}
 
 # Save the language code for HTTP requests and set the locale for l10n
 loc = locale.getdefaultlocale()[0]
@@ -363,9 +345,9 @@ def MainMenu():
     if False is not UsePrimeVideo:
         if 0 == len(pvCatalog):
             ''' Build the root catalog '''
-            if not PrimeVideo_BuildRoot():
+            if not g.pv.BuildRoot():
                 return
-        PrimeVideo_Browse('root')
+        g.pv.Browse('root')
     else:
         loadCategories()
 
@@ -392,622 +374,6 @@ def BeautifyText(title):
               (r'\.\.\.', '…')]: # Replace triple dots with ellipsis
         title = re.sub(t[0], t[1], title)
     return title
-
-
-def PrimeVideo_BuildRoot():
-    """ Parse the top menu on primevideo.com and build the root catalog """
-    pvCatalog['root'] = OrderedDict()
-    home = getURL(BaseUrl, silent=True, useCookie=True, rjson=False)
-    if None is home:
-        Dialog.notification('Connection error', 'Unable to fetch the primevideo.com homepage', xbmcgui.NOTIFICATION_ERROR)
-        Log('Unable to fetch the primevideo.com homepage', xbmc.LOGERROR)
-        return False
-    watchlist = re.search(r'<a[^>]* href="(/watchlist/)[^"]*"[^>]*>(.*?)</a>', home)
-    if None is not watchlist:
-        pvCatalog['root']['Watchlist'] = { 'title': watchlist.group(2), 'lazyLoadURL': BaseUrl + watchlist.group(1) }
-    home = re.search('<div id="av-nav-main-menu".*?<ul role="navigation"[^>]*>\s*(.*?)\s*</ul>', home)
-    if None is home:
-        Dialog.notification('PrimeVideo error', 'Unable to find the main primevideo.com navigation section', xbmcgui.NOTIFICATION_ERROR)
-        Log('Unable to find the main primevideo.com navigation section', xbmc.LOGERROR)
-        return False
-    for item in re.findall('<li[^>]*>\s*(.*?)\s*</li>', home.group(1)):
-        item = re.search('<a href="([^"]+)"[^>]*>\s*(.*?)\s*</a>', item)
-        if None is not re.match('/storefront/home/', item.group(1)):
-            continue
-        # Remove react comments
-        title = re.sub('^<!--\s*[^>]+\s*-->', '', re.sub('<!--\s*[^>]+\s*-->$', '', item.group(2)))
-        pvCatalog['root'][title] = {'title': title, 'lazyLoadURL': BaseUrl + item.group(1) + '?_encoding=UTF8&format=json'}
-    if 0 == len(pvCatalog['root']):
-        Dialog.notification('PrimeVideo error', 'Unable to build the root catalog from primevideo.com', xbmcgui.NOTIFICATION_ERROR)
-        Log('Unable to build the root catalog from primevideo.com', xbmc.LOGERROR)
-        return False
-    ''' Add functionalities to the root catalog '''
-    pvCatalog['root']['Search'] = {'title': getString(30108), 'verb': 'PrimeVideo_Search'}
-
-    # Set the expiration in 11 hours
-    pvCatalog['expiration'] = 39600 + int(time.time())
-
-    # Flush
-    with open(PrimeVideoCache, 'w+') as fp:
-        pickle.dump(pvCatalog, fp)
-
-    return True
-
-
-def PrimeVideo_Search():
-    """ Provide search functionality for PrimeVideo """
-    searchString = Dialog.input(getString(24121)).strip(' \t\n\r')
-    if 0 == len(searchString):
-        xbmcplugin.endOfDirectory(pluginhandle, succeeded=False)
-        return
-    Log('Searching "{0}"…'.format(searchString), xbmc.LOGINFO)
-    pvCatalog['search'] = OrderedDict([('lazyLoadURL', 'https://www.primevideo.com/search/ref=atv_nb_sr?ie=UTF8&phrase={0}'.format(searchString))])
-    PrimeVideo_Browse('search', xbmcplugin.SORT_METHOD_NONE)
-
-
-def PrimeVideo_Browse(path, forceSort=None):
-    """ Display and navigate the menu for PrimeVideo users """
-
-    path = path.decode('utf-8')
-
-    # Add multiuser menu if needed
-    if (multiuser) and ('root' == path) and (1 < len(loadUsers())):
-        li = xbmcgui.ListItem(getString(30134).format(loadUser('name')))
-        li.addContextMenuItems(ContextMenu_MultiUser())
-        xbmcplugin.addDirectoryItem(pluginhandle, '{0}?mode=PrimeVideo_Browse&path=root-//-SwitchUser'.format(sys.argv[0]), li, isFolder=False)
-    if 'root-//-SwitchUser' == path:
-        if switchUser():
-            PrimeVideo_BuildRoot()
-        return
-
-    node = pvCatalog
-    nodeName = None
-    for n in path.split('-//-'):
-        nodeName = n
-        node = node[n]
-
-    if (nodeName in pvVideoData) and ('metadata' in pvVideoData[nodeName]) and ('video' in pvVideoData[nodeName]['metadata']):
-        """ Start the playback if on a video leaf """
-        PlayVideo(pvVideoData[nodeName]['metadata']['video'], pvVideoData[nodeName]['metadata']['asin'], '', 0)
-        return
-
-    if (None is not node) and ('lazyLoadURL' in node):
-        PrimeVideo_LazyLoad(node, nodeName)
-
-    if None is node:
-        Dialog.notification('General error', 'Something went terribly wrong…', xbmcgui.NOTIFICATION_ERROR)
-        return
-
-    folderType = 0 if 'root' == path else 1
-    for key in node:
-        if key in ['metadata', 'ref', 'title', 'verb', 'children']:
-            continue
-        url = '{0}?mode='.format(sys.argv[0])
-        entry = node[key] if key not in pvVideoData else pvVideoData[key]
-        if 'verb' not in entry:
-            url += 'PrimeVideo_Browse&path={0}-//-{1}'.format(urllib.quote_plus(path.encode('utf-8')), urllib.quote_plus(key.encode('utf-8')))
-            # Squash season number folder when only one season is available
-            if ('metadata' not in entry) and ('children' in entry) and (1 == len(entry['children'])):
-                url += '-//-{0}'.format(urllib.quote_plus(entry['children'][0].encode('utf-8')))
-        else:
-            url += entry['verb']
-        title = entry['title'] if 'title' in entry else nodeName
-        item = xbmcgui.ListItem(title)
-        folder = True
-        # In case of series find the oldest series and apply its art, also update metadata in case of squashed series
-        if ('metadata' not in entry) and ('children' in entry) and (0 < len(entry['children'])):
-            if 1 == len(entry['children']):
-                entry['metadata'] = { 'artmeta': pvVideoData[entry['children'][0]]['metadata']['artmeta'], 'videometa': pvVideoData[entry['children'][0]]['metadata']['videometa'] }
-            else:
-                sn = 90001
-                snid = None
-                for child in entry['children']:
-                    if ('season' in pvVideoData[child]['metadata']['videometa']) and (sn > pvVideoData[child]['metadata']['videometa']['season']):
-                        sn = pvVideoData[child]['metadata']['videometa']['season']
-                        snid = child
-                if None is not snid:
-                    entry['metadata'] = { 'artmeta': pvVideoData[snid]['metadata']['artmeta'], 'videometa': { 'mediatype': 'tvshow' } }
-        if 'metadata' in entry:
-            m = entry['metadata']
-            if 'artmeta' in m: item.setArt(m['artmeta'])
-            if 'videometa' in m:
-                # https://codedocs.xyz/xbmc/xbmc/group__python__xbmcgui__listitem.html#ga0b71166869bda87ad744942888fb5f14
-                item.setInfo('video', m['videometa'])
-                if 'episode' in m['videometa']:
-                    folderType = 4  # Episode
-                elif 'season' in m['videometa']:
-                    if 'tvshow' == m['videometa']['mediatype']:
-                        folderType = 2  # Series list
-                    else:
-                        folderType = 3  # Season
-                elif 2 > folderType:  # If it's not been declared series, season or episode yet…
-                    folderType = 5  # … it's a Movie
-            if 'video' in m:
-                folder = False
-                item.setProperty('IsPlayable', 'true')
-                item.setInfo('video', {'title': title})
-                if 'runtime' in m:
-                    item.setInfo('video', {'duration': m['runtime']})
-                    item.addStreamInfo('video', {'duration': m['runtime']})
-        xbmcplugin.addDirectoryItem(pluginhandle, url, item, isFolder=folder)
-        del item
-
-    # Set sort method and view
-    # https://codedocs.xyz/xbmc/xbmc/group__python__xbmcplugin.html#ga85b3bff796fd644fb28f87b136025f40
-    xbmcplugin.addSortMethod(pluginhandle, [
-        xbmcplugin.SORT_METHOD_NONE,
-        xbmcplugin.SORT_METHOD_LABEL_IGNORE_THE,
-        xbmcplugin.SORT_METHOD_LABEL_IGNORE_THE,
-        xbmcplugin.SORT_METHOD_LABEL_IGNORE_THE,
-        xbmcplugin.SORT_METHOD_EPISODE,
-        xbmcplugin.SORT_METHOD_LABEL_IGNORE_THE,
-    ][folderType if None is forceSort else forceSort])
-
-    if 'false' == addon.getSetting("viewenable"):
-        # Only vfs and videos to keep Kodi's watched functionalities
-        folderType = 0 if 2 > folderType else 1
-    else:
-        # Actual views, set the main categories as vfs
-        folderType = 0 if 2 > folderType else 2
-
-    setContentAndView([None, 'videos', 'series', 'season', 'episode', 'movie'][folderType])
-
-
-def PrimeVideo_LazyLoad(obj, objName):
-    def Unescape(text):
-        """ Unescape various html/xml entities, courtesy of Fredrik Lundh """
-
-        def fixup(m):
-            import htmlentitydefs
-            text = m.group(0)
-            if text[:2] == "&#":
-                # character reference
-                try:
-                    if text[:3] == "&#x":
-                        return unichr(int(text[3:-1], 16))
-                    else:
-                        return unichr(int(text[2:-1]))
-                except ValueError:
-                    pass
-            else:
-                # named entity
-                try:
-                    text = unichr(htmlentitydefs.name2codepoint[text[1:-1]])
-                except KeyError:
-                    pass
-            return text  # leave as is
-
-        # Since we're using this for titles and synopses, and clean up general mess
-        ret = re.sub("&#?\w+;", fixup, text)
-        if ('"' == ret[0:1]) and ('"' == ret[-1:]):
-            ret = ret[1:-1]
-        
-        # Try to correct text when Amazon returns latin-1 encoded utf-8 characters
-        # (with the help of https://ftfy.now.sh/)
-        try:
-            ret = ret.encode('latin-1').decode('utf-8')
-        except (UnicodeEncodeError, UnicodeDecodeError):
-            pass
-
-        return BeautifyText(ret)
-
-    def MaxSize(imgUrl):
-        """ Strip the dynamic resize triggers from the URL """
-        return re.sub(r'_(SX[0-9]+|UR[0-9]+,[0-9]+)', '', imgUrl)
-
-    def ExtractURN(url):
-        """ Extract the unique resource name identifier """
-        ret = re.search(r'(/gp/video)?/detail/([^/]+)/', url)
-        if None is not ret:
-            ret = ret.group(2)
-        return ret
-
-    def DelocalizeDate(lang, datestr):
-        """ Convert language based timestamps into YYYY-MM-DD """
-        if lang not in dateParserData:
-            Log('Unable decode date "{0}": language "{1}" not supported'.format(datestr, lang), xbmc.LOGWARNING)
-            return datestr
-        p = re.search(dateParserData[lang]['deconstruct'], datestr)
-        if None is p:
-            # Sometimes Amazon returns english everything, let's try to figure out if this is the case
-            p = re.search(dateParserData['en_US']['deconstruct'], datestr)
-            if None is p:
-                Log('Unable to parse date "{0}" with language "{1}": format changed?'.format(datestr, lang), xbmc.LOGWARNING)
-                return datestr
-        p = list(p.groups())
-        p[dateParserData[lang]['month']] = dateParserData[lang]['months'][p[dateParserData[lang]['month']]]
-        return dateParserData[lang]['reassemble'].format(p[0], p[1], p[2])
-
-    def NotifyUser(msg):
-        """ Pop up messages while scraping to inform users of progress """
-        if not hasattr(NotifyUser, 'lastNotification'):
-            NotifyUser.lastNotification = 0
-        if NotifyUser.lastNotification < time.time():
-            ''' Only update once every other second, to avoid endless message queue '''
-            NotifyUser.lastNotification = 1 + time.time()
-            Dialog.notification(__plugin__, msg, time=1000, sound=False)
-
-    def PopulateInlineMovieSeries(obj, rq, title, image, url):
-        """ Takes a sequence of Movies or TV Series such as in watchlist and carousels and fetches the metadata """
-        Log('PIMS: %s %s %s %s %s' % (obj, rq, title, image, url))
-        urn = ExtractURN(url)
-        obj[urn] = {}
-        if urn in pvVideoData:
-            return False
-        pvVideoData[urn] = {
-        'metadata': {'artmeta': {'thumb': MaxSize(image), 'poster': MaxSize(image)}, 'videometa': {}},
-        'title': Unescape(title)}
-        rq.append((url, obj[urn], urn))
-        return True
-
-    # Set up the fetch order to find the best quality image possible
-    imageSizes = r'(large-screen-double|desktop-double|tablet-landscape-double|phone-landscape-double|phablet-double|phone-double|large-screen|desktop|tablet-landscape|tablet|phone-landscape|phablet|phone)'
-
-    if 'lazyLoadURL' not in obj:
-        return
-    requestURLs = [obj['lazyLoadURL']]
-
-    amzLang = None
-    if None is not requestURLs[0]:
-        # Fine the locale amazon's using
-        cj = MechanizeLogin()
-        if cj:
-            amzLang = cj.get('lc-main-av', domain='.primevideo.com', path='/')
-
-    bUpdatedVideoData = False  # Whether or not the pvData has been updated
-
-    while 0 < len(requestURLs):
-        # We linearize and avoid recursion through requestURLs differentiation
-        # - String: request url, the obj parameter is the root object
-        # - Tuple(URL, object, [urn]): request url, object reference and its urn
-        if not isinstance(requestURLs[0], tuple):
-            requestURL = requestURLs[0]
-            o = obj
-            refUrn = objName
-        else:
-            requestURL = requestURLs[0][0]
-            o = requestURLs[0][1]
-            refUrn = requestURLs[0][2]
-        del requestURLs[0]
-        couldNotParse = False
-        try:
-            cnt = getURL(requestURL, silent=True, useCookie=True, rjson=False)
-            if 'lazyLoadURL' in o:
-                o['ref'] = o['lazyLoadURL']
-                del o['lazyLoadURL']
-        except:
-            couldNotParse = True
-        if couldNotParse or (0 == len(cnt)):
-            Dialog.notification(getString(30251), requestURL, xbmcgui.NOTIFICATION_ERROR)
-            Log('Unable to fetch the url: {0}'.format(requestURL), xbmc.LOGERROR)
-            break
-
-        for t in [('\\\\n', '\n'), ('\\n', '\n'), ('\\\\"', '"'), (r'^\s+', '')]:
-            cnt = re.sub(t[0], t[1], cnt, flags=re.DOTALL)
-        if None is not re.search('<html[^>]*>', cnt):
-            ''' If there's an HTML tag it's no JSON-AmazonUI-Streaming object '''
-            if None is not re.search(r'<div[^>]* id="Watchlist"[^>]*>', cnt, flags=re.DOTALL):
-                ''' Watchlist '''
-                if not hasattr(PrimeVideo_LazyLoad, 'WatchlistPages'):
-                    PrimeVideo_LazyLoad.WatchlistPages = {}
-                if ('Watchlist' == objName):
-                    for entry in re.findall(r'<a href="([^"]+/)[^"/]+" class="DigitalVideoUI_TabHeading__tab([^"]*DigitalVideoUI_TabHeading__active)?">(.*?)</a>', cnt, flags=re.DOTALL):
-                        obj[Unescape(entry[2])] = { 'title': Unescape(entry[2]), 'lazyLoadURL': BaseUrl + entry[0] + '?sort=DATE_ADDED_DESC' }
-                    continue
-                for entry in re.findall(r'<div[^>]* class="[^"]*DigitalVideoWebNodeLists_Item__item[^"]*"[^>]*>\s*<a href="(/detail/[^/]+/)[^"]*"[^>]*>*.*?<img src="([^"]+)".*?"[^"]*DigitalVideoWebNodeLists_Item__coreTitle[^"]*"[^>]*>\s*(.*?)\s*</', cnt):
-                    bUpdatedVideoData = True if PopulateInlineMovieSeries(obj, requestURLs, entry[2], entry[1], BaseUrl + entry[0]) else bUpdatedVideoData
-                pagination = re.search(r'<ol[^>]* class="[^"]*DigitalVideoUI_Pagination__pagination[^"]*"[^>]*>(.*?)</ol>', cnt)
-                if None is not pagination:
-                    # We save a semi-static list of scraped pages, to avoid circular loops
-                    if (refUrn not in PrimeVideo_LazyLoad.WatchlistPages):
-                        PrimeVideo_LazyLoad.WatchlistPages[refUrn] = []
-                    currentPage = re.search(r'<a[^>]* href="#"[^>]*>\s*<span>\s*([0-9]+)\s*</span>', pagination.group(0), flags=re.DOTALL).group(1)
-                    if currentPage not in PrimeVideo_LazyLoad.WatchlistPages[refUrn]:
-                        PrimeVideo_LazyLoad.WatchlistPages[refUrn].append(currentPage)
-                    for entry in re.findall(r'<li[^>]*>\s*<a href="(/[^"]+)"[^>]*>\s*<span>\s*([0-9]+)\s*', pagination.group(0)):
-                        if entry[1] not in PrimeVideo_LazyLoad.WatchlistPages[refUrn]:
-                            requestURLs.append((BaseUrl + entry[0], o, refUrn))
-            elif None is re.search(r'"[^"]*av-result-cards[^"]*"', cnt, flags=re.DOTALL):
-                ''' Movie/series page '''
-                # Find the biggest fanart available
-                bgimg = re.search(r'<div class="av-hero-background-size av-bgimg av-bgimg-' + imageSizes + r'">.*?url\(([^)]+)\)', cnt, flags=re.DOTALL)
-                if None is not bgimg:
-                    bgimg = MaxSize(bgimg.group(2))
-
-                # Extract the global data
-                rx = [
-                    r'<span data-automation-id="imdb-release-year-badge"[^>]*>\s*([0-9]+)\s*</span>',  # Year
-                    r'<span data-automation-id="imdb-rating-badge"[^>]*>\s*([0-9]+)[,.]([0-9]+)\s*</span>',  # IMDb rating
-                    r'<span data-automation-id="maturity-rating-badge"[^>]*>\s*(.*?)\s*</span>',  # Age rating
-                    r'<dt data-automation-id="meta-info-starring"[^>]*>[^<]*</dt>\s*<dd[^>]*>\s*(.*?)\s*</dd>',  # Starring
-                    r'<dt data-automation-id="meta-info-genres"[^>]*>[^<]*</dt>\s*<dd[^>]*>\s*(.*?)\s*</dd>',  # Genre
-                    r'<dt data-automation-id="meta-info-director"[^>]*>[^<]*</dt>\s*<dd[^>]*>\s*(.*?)\s*</dd>',  # Director
-                    r'<div data-automation-id="synopsis"[^>]*>[^<]*<p>\s*(.*?)\s*</p>',  # Synopsis
-                ]
-                results = re.search(r'<section\s+[^>]*class="[^"]*av-detail-section[^"]*"[^>]*>\s*(.*?)\s*</section>', cnt, flags=re.DOTALL).group(1)
-                gres = []
-                for i in range(0, len(rx)):
-                    gres.append(re.search(rx[i], results, flags=re.DOTALL))
-                    if None is not gres[i]:
-                        gres[i] = gres[i].groups()
-                        if 1 == len(gres[i]):
-                            gres[i] = Unescape(gres[i][0])
-                        if (2 < i) and (6 != i):
-                            gres[i] = re.sub(r'\s*</?a[^>]*>\s*', '', gres[i])
-                            gres[i] = re.split(r'\s*[,;]\s*', gres[i])
-                            # Cast is always to be sent as a list, single string is only required/preferred for Genre and Director
-                            if (3 < i) and (1 == len(gres[i])):
-                                gres[i] = gres[i][0]
-
-                results = re.search(r'<ol\s+[^>]*class="[^"]*av-episode-list[^"]*"[^>]*>\s*(.*?)\s*</ol>', cnt, flags=re.DOTALL)
-                if None is results:
-                    ''' Standalone movie '''
-                    if 'video' not in pvVideoData[refUrn]['metadata']:
-                        asin = re.search(r'"asin":"([^"]*)"', cnt, flags=re.DOTALL)
-                        if None is not asin:
-                            bUpdatedVideoData = True
-
-                            meta = pvVideoData[refUrn]['metadata']
-                            if None is not bgimg:
-                                meta['artmeta']['fanart'] = bgimg
-                            NotifyUser(getString(30253).format(pvVideoData[refUrn]['title']))
-                            meta['asin'] = asin.group(1)
-                            meta['video'] = ExtractURN(requestURL)
-
-                            # Insert movie information
-                            if None is not gres[0]: meta['videometa']['year'] = gres[0]
-                            if None is not gres[1]: meta['videometa']['rating'] = int(gres[1][0]) + (int(gres[1][1]) / 10.0)
-                            if None is not gres[2]: meta['videometa']['mpaa'] = gres[2]
-                            if None is not gres[3]: meta['videometa']['cast'] = gres[3]
-                            if None is not gres[4]: meta['videometa']['genre'] = gres[4]
-                            if None is not gres[5]: meta['videometa']['director'] = gres[5]
-                            if None is not gres[6]: meta['videometa']['plot'] = gres[6]
-
-                            # Extract the runtime
-                            success, gpr = getUrldata('catalog/GetPlaybackResources', meta['asin'], useCookie=True, extra=True,
-                                                    opt='&titleDecorationScheme=primary-content', dRes='CatalogMetadata')
-                            if not success:
-                                gpr = None
-                            else:
-                                if 'runtimeSeconds' in gpr['catalogMetadata']['catalog']:
-                                    meta['runtime'] = gpr['catalogMetadata']['catalog']['runtimeSeconds']
-                else:
-                    ''' Episode list '''
-                    for entry in re.findall(r'<li[^>]*>\s*(.*?)\s*</li>', results.group(1), flags=re.DOTALL):
-                        rx = [
-                            r'<a data-automation-id="ep-playback-[0-9]+"[^>]*data-ref="([^"]*)"[^>]*data-title-id="([^"]*)"[^>]*href="([^"]*)"',
-                            r'<div class="av-bgimg av-bgimg-' + imageSizes + r'">.*?url\(([^)]+)\)',  # Image
-                            r'<span class="av-play-title-text">\s*(.*?)\s*</span>',  # Title
-                            r'<span data-automation-id="ep-air-date-badge-[0-9]+"[^>]*>\s*(.*?)\s*</span>',  # Original air date
-                            r'<span data-automation-id="ep-amr-badge-[0-9]+"[^>]*>\s*(.*?)\s*</span>',  # Age rating
-                            r'<p data-automation-id="ep-synopsis-[0-9]+"[^>]*>\s*(.*?)\s*</p>',  # Synopsis
-                            r'id="ep-playback-([0-9]+)"',  # Episode number
-                            r'\s+data-title-id="\s*([^"]+?)\s*"',  # Episode's asin
-                        ]
-                        res = []
-                        for i in range(0, len(rx)):
-                            res.append(re.search(rx[i], entry))
-                            if None is not res[i]:
-                                res[i] = res[i].groups()
-                                # If holding a single result, don't provide a list
-                                if 1 == len(res[i]):
-                                    res[i] = Unescape(res[i][0])
-                                # We just need the image, not the type                            
-                                if 1 == i:
-                                    res[i] = res[i][1]
-                        if (None is res[0]) or (None is res[2]):
-                            ''' Some episodes might be not playable until a later date or just N/A, although listed '''
-                            continue
-                        meta = {'artmeta': {'thumb': MaxSize(res[1]), 'poster': MaxSize(res[1]), 'fanart': bgimg, }, 'videometa': {'mediatype': 'episode'},
-                                'id': res[0][0], 'asin': res[0][1], 'videoURL': res[0][2]}
-                        if None is not re.match(r'/[^/]', meta['videoURL']):
-                            meta['videoURL'] = BaseUrl + meta['videoURL']
-                        meta['video'] = ExtractURN(meta['videoURL'])
-                        eid = meta['video']
-                        #if None is not res[7]:
-                        #    eid = res[7]
-
-                        if (eid not in pvVideoData) or ('videometa' not in pvVideoData[eid]) or (0 == len(pvVideoData[eid]['videometa'])):
-                            # Extract the runtime
-                            success, gpr = getUrldata('catalog/GetPlaybackResources', res[7], useCookie=True, extra=True,
-                                                    opt='&titleDecorationScheme=primary-content', dRes='CatalogMetadata')
-                            if not success:
-                                gpr = None
-                            else:
-                                if 'runtimeSeconds' in gpr['catalogMetadata']['catalog']:
-                                    meta['runtime'] = gpr['catalogMetadata']['catalog']['runtimeSeconds']
-
-                            # Sometimes the chicken comes before the egg
-                            #if refUrn not in pvVideoData:
-                            #    pvVideoData[refUrn] = { 'metadata': { 'videometa': {} }, 'children': [] }
-
-                            # Insert series information
-                            if None is not gres[0]: meta['videometa']['year'] = gres[0]
-                            if None is not gres[1]: meta['videometa']['rating'] = int(gres[1][0]) + (int(gres[1][1]) / 10.0)
-                            if None is not gres[2]: meta['videometa']['mpaa'] = gres[2]
-                            if None is not gres[3]:
-                                meta['videometa']['cast'] = gres[3]
-                                pvVideoData[refUrn]['metadata']['videometa']['cast'] = gres[3]
-                            if None is not gres[4]:
-                                meta['videometa']['genre'] = gres[4]
-                                pvVideoData[refUrn]['metadata']['videometa']['genre'] = gres[4]
-                            if None is not gres[5]:
-                                meta['videometa']['director'] = gres[5]
-                                pvVideoData[refUrn]['metadata']['videometa']['director'] = gres[5]
-
-                            # Insert episode specific information
-                            if None is not res[3]: meta['videometa']['premiered'] = DelocalizeDate(amzLang, res[3])
-                            if None is not res[4]: meta['videometa']['mpaa'] = res[4]
-                            if None is not res[5]: meta['videometa']['plot'] = res[5]
-                            if None is not res[6]:
-                                if 'season' in pvVideoData[refUrn]['metadata']['videometa']:
-                                    meta['videometa']['season'] = pvVideoData[refUrn]['metadata']['videometa']['season']
-                                else:
-                                    ''' We're coming from a widow carousel '''
-                                    pvVideoData[refUrn]['metadata']['videometa']['mediatype'] = 'tvshow'
-                                    # Try with page season selector
-                                    ssn = re.search(r'<a class="av-droplist--selected"[^>]*>[^0-9]*([0-9]+)[^0-9]*</a>', cnt)
-                                    if None is ssn:
-                                        # Try with single season
-                                        ssn = re.search(r'<span class="av-season-single av-badge-text">[^0-9]*([0-9]+)[^0-9]*</span>', cnt)
-                                    if None is not ssn:
-                                        ssn = ssn.group(1)
-                                        meta['videometa']['season'] = ssn
-                                        pvVideoData[refUrn]['metadata']['videometa']['season'] = ssn
-                                meta['videometa']['episode'] = int(res[6])
-
-                            # Episode title cleanup
-                            title = res[2]
-                            if None is not re.match(r'[0-9]+[.]\s*', title):
-                                ''' Strip the episode number '''
-                                title = re.sub(r'^[0-9]+.\s*', '', title)
-                            else:
-                                ''' Probably an extra trailer or something, remove episode information '''
-                                del meta['videometa']['season']
-                                del meta['videometa']['episode']
-
-                            bUpdatedVideoData = True
-                            pvVideoData[eid] = { 'metadata': meta, 'title': title, 'parent': refUrn }
-                            NotifyUser(getString(30253).format(title))
-                        if eid not in o:
-                            o[eid] = {}
-                        if (refUrn in pvVideoData) and ('children' in pvVideoData[refUrn]) and (eid not in pvVideoData[refUrn]['children']):
-                            pvVideoData[refUrn]['children'].append(eid)
-
-            else:
-                ''' Movie and series list '''
-                results = re.search(r'<ol\s+[^>]*class="[^"]*av-result-cards[^"]*"[^>]*>\s*(.*?)\s*</ol>', cnt, flags=re.DOTALL)
-                for entry in re.findall(r'<li[^>]*>\s*(.*?)\s*</li>', results.group(1), flags=re.DOTALL):
-                    rx = [
-                        r'<img\s+[^>]*src="([^"]*)".*?<h2[^>]*>\s*<a\s+[^>]*href="([^"]+)"[^>]*>\s*(.*?)\s*</a>\s*</h2>',  # Image, link and title
-                        r'<span\s+[^>]*class="[^"]*av-result-card-rating[^"]*"[^>]*>\s*([0-9]+)[,.]([0-9]+)\s*</span>',  # IMDb rating
-                        r'<p\s+[^>]*class="[^"]*av-result-card-synopsis[^"]*"[^>]*>\s*(.*?)\s*</p>',  # Synopsis
-                        r'<span\s+[^>]*class="[^"]*av-result-card-season[^"]*"[^>]*>\s*(.*?)\s*</span>',  # Season
-                        r'<span\s+[^>]*class="[^"]*av-result-card-year[^"]*"[^>]*>\s*(.*?)\s*</span>',  # Year
-                        r'<span\s+[^>]*class="[^"]*av-result-card-certification[^"]*"[^>]*>\s*(.*?)\s*</span>',  # Age rating
-                        r'\s+data-asin="\s*([^"]+?)\s*"',  # Movie/episode's URN
-                    ]
-                    res = []
-                    for i in range(0, len(rx)):
-                        res.append(re.search(rx[i], entry))
-                        if None is not res[i]:
-                            res[i] = res[i].groups()
-                            if 1 == len(res[i]):
-                                res[i] = Unescape(res[i][0])
-                            else:
-                                res[i] = list(res[i])
-                    title = Unescape(res[0][2])
-                    NotifyUser(getString(30253).format(title))
-                    if None is not re.match(r'/[^/]', res[0][1]):
-                        res[0][1] = BaseUrl + res[0][1]
-                    meta = {'artmeta': {'thumb': MaxSize(res[0][0]), 'poster': MaxSize(res[0][0])}, 'videometa': {}}
-                    if None is not res[1]: meta['videometa']['rating'] = int(res[1][0]) + (int(res[1][1]) / 10.0)
-                    if None is not res[2]: meta['videometa']['plot'] = res[2]
-                    meta['videometa']['mediatype'] = 'movie' if None is res[3] else 'season'
-                    if None is not res[4]: meta['videometa']['year'] = int(res[4])
-                    if None is not res[5]: meta['videometa']['mpaa'] = res[5]
-
-                    if None is res[3]:
-                        ''' Movie '''
-                        # Queue the movie page instead of parsing partial information here
-                        urn = ExtractURN(res[0][1])
-                        if urn not in o:
-                            o[urn] = {}
-                        if (urn not in pvVideoData) or ('metadata' not in pvVideoData[urn]) or ('video' not in pvVideoData[urn]['metadata']):
-                            bUpdatedVideoData = True
-                            pvVideoData[urn] = {'metadata': meta, 'title': title}
-                            requestURLs.append((res[0][1], o[urn], urn))
-                    else:
-                        ''' Series '''
-                        id = title
-                        sid = ExtractURN(res[0][1])
-
-                        if sid not in pvVideoData:
-                            pvVideoData[sid] = { 'children': [] }
-
-                        # Extract video metadata
-                        if 'parent' in pvVideoData[sid]:
-                            id = pvVideoData[sid]['parent']
-                        elif (id is title) and (None is not res[6]):
-                            success, gpr = getUrldata('catalog/GetPlaybackResources', res[6], useCookie=True, extra=True,
-                                                      opt='&titleDecorationScheme=primary-content', dRes='CatalogMetadata')
-                            if not success:
-                                Log('Unable to get the video metadata for {0} ({1})'.format(title, res[0][1]), xbmc.LOGWARNING)
-                            else:
-                                # Find the show Asin URN, if possible
-                                for d in gpr['catalogMetadata']['family']['tvAncestors']:
-                                    if 'SHOW' == d['catalog']['type']:
-                                        id = d['catalog']['id']
-                                        pvVideoData[sid]['parent'] = id
-                                        break
-                        if id not in o:
-                            o[id] = { 'title': title }
-                        if (sid in pvVideoData) and ('children' in pvVideoData[sid]) and (0 < len(pvVideoData[sid]['children'])):
-                            o[id][sid] = { k: {} for k in pvVideoData[sid]['children'] }
-                        else:
-                            # Save tree information if id is a URN and not a title name
-                            if (id is not title):
-                                if id not in pvVideoData:
-                                    pvVideoData[id] = { 'children': [] }
-                                if sid not in pvVideoData[id]['children']:
-                                    pvVideoData[id]['children'].append(sid)
-                                if 'title' not in pvVideoData[id]:
-                                    pvVideoData[id]['title'] = title
-                            sn = res[3]
-                            n = int(re.sub(r'^[^0-9]*([0-9]+)[^0-9]*$', r'\1', sn))
-                            meta['videometa']['season'] = n
-                            o[id][sid] = { 'lazyLoadURL': res[0][1] }
-                            bUpdatedVideoData = True
-                            pvVideoData[sid]['title'] = sn
-                            pvVideoData[sid]['metadata'] = meta
-
-                # Next page
-                pagination = re.search(
-                    r'<ol\s+[^>]*id="[^"]*av-pagination[^"]*"[^>]*>.*?<li\s+[^>]*class="[^"]*av-pagination-current-page[^"]*"[^>]*>.*?</li>\s*<li\s+[^>]*class="av-pagination[^>]*>\s*(.*?)\s*</li>\s*</ol>',
-                    cnt, flags=re.DOTALL)
-                if None is not pagination:
-                    nru = Unescape(re.search(r'href="([^"]+)"', pagination.group(1), flags=re.DOTALL).group(1))
-                    if None is not re.match(r'/[^/]', nru):
-                        nru = BaseUrl + nru
-                    requestURLs.append(nru)
-        else:
-            ''' Categories list '''
-            for section in re.split(r'&&&\s+', cnt):
-                if 0 == len(section):
-                    continue
-                section = re.split(r'","', section[2:-2])
-                if 'dvappend' == section[0]:
-                    title = Unescape(re.sub(r'^.*<h2[^>]*>\s*<span[^>]*>\s*(.*?)\s*</span>.*$', r'\1', section[2], flags=re.DOTALL))
-                    NotifyUser(getString(30253).format(title))
-                    o[title] = {'title': title}
-                    if None is not re.search('<h2[^>]*>.*?<a\s+[^>]*\s+href="[^"]+"[^>]*>.*?</h2>', section[2], flags=re.DOTALL):
-                        o[title]['lazyLoadURL'] = Unescape(
-                            re.sub('\\n', '', re.sub(r'^.*?<h2[^>]*>.*?<a\s+[^>]*\s+href="([^"]+)"[^>]*>.*?</h2>.*?$', r'\1', section[2], flags=re.DOTALL)))
-                    else:
-                        ''' The carousel has no explore link, we need to parse what we can from the carousel itself '''
-                        for entry in re.findall(r'<li[^>]*>\s*(.*?)\s*</li>', section[2], flags=re.DOTALL):
-                            parts = re.search(
-                                r'<a\s+[^>]*href="([^"]*)"[^>]*>\s*.*?(src|data-a-image-source)="([^"]*)"[^>]*>.*?class="dv-core-title"[^>]*>\s*(.*?)\s*</span>',
-                                entry, flags=re.DOTALL)
-                            if None is not re.search(r'/search/', parts.group(1)):
-                                ''' Category '''
-                                o[title][parts.group(4)] = {'metadata': {'artmeta': {'thumb': MaxSize(parts.group(3)), 'poster': MaxSize(parts.group(3))}},
-                                                            'lazyLoadURL': parts.group(1), 'title': parts.group(4)}
-                            else:
-                                ''' Movie/Series list '''
-                                bUpdatedVideoData = True if PopulateInlineMovieSeries(o[title], requestURLs, parts.group(4), parts.group(3), parts.group(1)) else bUpdatedVideoData
-                pagination = re.search(r'data-ajax-pagination="{&quot;href&quot;:&quot;([^}]+)&quot;}"', section[2], flags=re.DOTALL)
-                if ('dvupdate' == section[0]) and (None is not pagination):
-                    requestURLs.append(re.sub(r'(&quot;,&quot;|&amp;)', '&', re.sub('&quot;:&quot;', '=', pagination.group(1) + '&format=json')))
-        if 0 < len(requestURLs):
-            NotifyUser(getString(30252))
-
-    # Flush catalog and data
-    with open(PrimeVideoCache, 'w+') as fp:
-        pickle.dump(pvCatalog, fp)
-    if bUpdatedVideoData:
-        with open(PrimeVideoData, 'w+') as fp:
-            pickle.dump(pvVideoData, fp)
 
 
 def Search():
@@ -3259,101 +2625,6 @@ class Captcha(pyxbmct.AddonDialogWindow):
         self.close()
 
 
-class AmazonPlayer(xbmc.Player):
-    def __init__(self):
-        super(AmazonPlayer, self).__init__()
-        self.sleeptm = 0.2
-        self.video_lastpos = 0
-        self.video_totaltime = 0
-        self.dbid = 0
-        self.seek = 0
-        self.url = ''
-        self.extern = ''
-        self.asin = ''
-        self.cookie = None
-        self.interval = 180
-        self.running = False
-
-    def resolve(self, li):
-        if not self.checkResume():
-            xbmcplugin.setResolvedUrl(pluginhandle, True, xbmcgui.ListItem())
-            xbmc.executebuiltin('Container.Refresh')
-            return
-        self.running = True
-        xbmcplugin.setResolvedUrl(pluginhandle, True, li)
-        self.PlayerInfo('Starting Playback')
-        if self.seek:
-            self.seekTime(self.seek)
-            self.PlayerInfo('Resuming Playback')
-
-        self.updateStream('START')
-        Log('Video ContentType Movie? %s' % xbmc.getCondVisibility('VideoPlayer.Content(movies)'), xbmc.LOGDEBUG)
-        Log('Video ContentType Episode? %s' % xbmc.getCondVisibility('VideoPlayer.Content(episodes)'), xbmc.LOGDEBUG)
-
-    def checkResume(self):
-        self.dbid = int('0' + getListItem('DBID'))
-        Log(self.dbid, xbmc.LOGDEBUG)
-        if not self.dbid:
-            return True
-        dbtype = getListItem('DBTYPE')
-        result = jsonRPC('VideoLibrary.Get%sDetails' % dbtype, 'resume,playcount', {'%sid' % dbtype: self.dbid})
-        position = int(result['episodedetails']['resume']['position'])
-        playcount = int(result['episodedetails']['playcount'])
-        Log(result, xbmc.LOGDEBUG)
-
-        if playcount:
-            return True
-        if position > 180:
-            sel = Dialog.contextmenu([getString(12022).format(time.strftime("%H:%M:%S", time.gmtime(position))), getString(12021)])
-            if sel > -1:
-                self.seek = position if sel == 0 else 0
-            else:
-                return False
-        return True
-
-    def onPlayBackEnded(self):
-        self.finished()
-
-    def onPlayBackStopped(self):
-        self.finished()
-
-    def updateStream(self, event):
-        suc, msg = getUrldata('usage/UpdateStream', self.asin, useCookie=self.cookie, opt='&event=%s&timecode=%s' % (event, self.video_lastpos))
-        if suc and 'statusCallbackIntervalSeconds' in str(msg):
-            self.interval = msg['message']['body']['statusCallbackIntervalSeconds']
-
-    def finished(self):
-        self.updateStream('STOP')
-        if self.extern:
-            playcount = 1 if (self.video_lastpos * 100) / self.video_totaltime >= 90 else 0
-            watched = getListItem('PlayCount')
-            if self.dbid:
-                dbtype = getListItem('DBTYPE')
-                params = {'%sid' % dbtype: self.dbid,
-                          'resume': {'position': 0 if playcount else self.video_lastpos,
-                                     'total': self.video_totaltime},
-                          'playcount': playcount}
-                res = '' if 'OK' in jsonRPC('VideoLibrary.Set%sDetails' % dbtype, '', params) else 'NOT '
-                Log('%sUpdated %sid(%s) with: pos(%s) total(%s) playcount(%s)' % (res, dbtype, self.dbid, self.video_lastpos, self.video_totaltime, playcount))
-            else:
-                Log('No DBID returned')
-                if playcount and not watched:
-                    xbmc.executebuiltin("Action(ToggleWatched)")
-
-        xbmc.executebuiltin('Container.Refresh')
-        self.running = False
-
-    def PlayerInfo(self, msg):
-        while not self.isPlayingVideo():
-            sleep(self.sleeptm)
-        while self.isPlayingVideo() and (0 > self.getTime() >= self.getTotalTime()):
-            sleep(self.sleeptm)
-        if self.isPlayingVideo():
-            self.video_totaltime = self.getTotalTime()
-            self.video_lastpos = self.getTime()
-            Log('%s: %s/%s' % (msg, self.video_lastpos, self.video_totaltime))
-
-
 args = dict(urlparse.parse_qsl(urlparse.urlparse(sys.argv[2]).query))
 
 Log(args, xbmc.LOGDEBUG)
@@ -3432,6 +2703,6 @@ elif mode == 'ageSettings':
     if RequestPin():
         AgeSettings(getString(30018).split('…')[0]).doModal()
 elif mode == 'PrimeVideo_Browse':
-    PrimeVideo_Browse(None if 'path' not in args else args['path'])
+    g.pv.Browse(None if 'path' not in args else args['path'])
 else:
     exec mode + '()'
